requirements.txt
# ============================================
# BarberOS - AI Receptionist Infrastructure
# ============================================

# --- Core Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# --- LangChain / LangGraph (Anti-Hallucination Agent) ---
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
langgraph==0.2.60
langsmith==0.2.10

# --- Vector Store / RAG ---
chromadb==0.5.23
tiktoken==0.8.0

# --- Web Scraping (CashBarber / AppBarber Integration) ---
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.49.1

# --- Database ---
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.1

# --- Observability & Monitoring ---
structlog==24.4.0
sentry-sdk[fastapi]==2.19.2
prometheus-client==0.21.1

# --- Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- Utilities ---
tenacity==9.0.0
cachetools==5.5.1
python-dateutil==2.9.0
PyYAML==6.0.2
Jinja2==3.1.4
redis==5.2.1

# --- Testing ---
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1
factory-boy==3.3.1
